requirements.txt
firebase-admin==6.5.0
pyrebase4==4.7.1
customtkinter==5.2.1
fpdf==1.7.2
requests==2.32.3
pillow==10.2.0
matplotlib==3.9.2

